N and prezto updates

+++ prezto.sh
@@ -2,12 +2,14 @@
 
 git submodule init
 git submodule update --recursive
+cd prezto && git submodule init && git submodule update --recursive && cd ..
 
 for rcfile in "${ZDOTDIR:-$HOME}"/.dotfiles/prezto/runcoms/*; do
     if ! echo $rcfile | grep "README.md"; then
-        echo "${ZDOTDIR:-$HOME}/.$(basename $rcfile)"
         ln -s "$rcfile" "${ZDOTDIR:-$HOME}/.$(basename $rcfile)"
     fi
 done
 
-ln -s $HOME/.dotfiles/prezto $HOME/.zprezto
+if [ ! -e $HOME/.zprezto ]; then
+    ln -s $HOME/.dotfiles/prezto $HOME/.zprezto
+fi
